backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
mysqlclient>=2.2,<3.0
dj-database-url>=2.1,<3.0

# Celery & Redis
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# Image Processing
Pillow>=10.0,<11.0

# Barcode Generation
python-barcode>=0.15,<1.0

# Report Generation
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Utilities
python-decouple>=3.8,<4.0
django-extensions>=3.2,<4.0

# Development & Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0

# Code Quality
ruff>=0.1,<1.0
black>=23.0,<24.0
isort>=5.12,<6.0
